requirements.txt
alabaster==0.7.10
Babel==2.4.0
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
Django==1.11.2
django-crispy-forms==1.6.1
django-multiselectfield==0.1.7
dnspython==1.15.0
docutils==0.14
email-validator==1.0.2
factory-boy==2.9.2
Faker==0.8.3
idna==2.6
imagesize==0.7.1
ipaddress==1.0.18
Jinja2==2.9.6
MarkupSafe==1.0
numpydoc==0.7.0
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
typing==3.6.2
urllib3==1.22
